fix: connect wallet padding & step folder moved
